+++ src/py/flwr/client/connection/grpc_rere/grpc_adapter_test.py
